pyproject.toml
[project]
name = "mcp-google-server"
version = "0.1.0"
description = "MCP server exposing Gmail and Google Calendar read tools over SSE"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "mcp>=1.2.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.30.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.130.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
